requirements.txt
Django>=3.0,<4.0
psycopg2-binary
djangorestframework==3.12.4
drf_yasg==1.20.0